requirements.txt
azure-functions
requests
beautifulsoup4
pandas
python-dotenv
sqlalchemy
pyodbc